+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Changelog for the docker image are [here](/docker_CHANGELOG.md).
 
 ## [Unreleased]
+- support offset in QueryDataOptions for loadCustomData/loadCustomEnts
 
 ## [0.2.2] - 2025-02-22
 - ability to customize default ordering for an indexed edge (#1878)